Replace numpy.float and np.float by float

+++ tslearn/tests/test_serialize_models.py
@@ -39,7 +39,7 @@ def clear_tmp():
 
 def test_hdftools():
     dtypes = [int, numpy.int8, numpy.int16, numpy.int32, numpy.int64,
-              numpy.float, numpy.float32, numpy.float64]
+              float, numpy.float32, numpy.float64]
 
     d = {}
